+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/factory/NDArrayFactory.java
@@ -257,7 +257,7 @@ public interface NDArrayFactory {
      * @param n the number for the identity
      * @return
      */
-    INDArray eye(int n);
+    INDArray eye(long n);
 
     /**
      * Rotate a matrix 90 degrees
